+++ controllers/tutorials.js
@@ -5,14 +5,12 @@ const prisma = new PrismaClient();
 const getAllTutorials = async (req, res) => {
     try {
         const tutorials = await prisma.tutorials.findMany();
-        res.status(200).json({
-            message: 'Get data success',
+        return res.status(200).json({
             data: tutorials
         })
     } catch (error) {
-        res.status(500).json({
-            message: 'Internal server error',
-            serverMessage: error
+        return res.status(500).json({
+            "error": `${error}`
         })
     }
 }
@@ -24,37 +22,35 @@ const getTutorialById = async (req, res) => {
             where: {id: Number(id)}
         })
         if(!tutorials){
-            return res.status(404).json({error: 'Data not found'})
+            return res.status(404).json({message: 'Data Not Found'})
         }
-        res.status(200).json({
-            message: 'Get data success',
+        return res.status(200).json({
             data: tutorials
         })
     } catch (error) {
-        res.status(500).json({
-            message: 'Internal server error',
-            serverMessage: error
+        return res.status(500).json({
+            "error": `${error}`
         })
     }
 }
 
 const createTutorial = async (req, res) => {
     try {
-        const {image_id, link} = req.body
+        const {link} = req.body
+        let {image_id} = req.body
+        image_id = Number(image_id)
         const tutorials = await prisma.tutorials.create({
             data: {
                 image_id: image_id,
                 link: link
             }
         })
-        res.status(201).json({
-            message: 'Add data success',
+        return res.status(201).json({
             data: tutorials
         })
     } catch (error) {
-        res.status(500).json({
-            message: 'Internal server error',
-            serverMessage: error
+        return res.status(500).json({
+            "error": `${error}`
         })
     }
 }
@@ -62,19 +58,22 @@ const createTutorial = async (req, res) => {
 const updateTutorial = async (req, res) => {
     try {
         const {id} = req.params
-        const {body} = req
+        const {link} = req.body
+        let {image_id} =req.body
+        image_id = Number(image_id)
         const tutorials = await prisma.tutorials.update({
             where: {id: Number(id)},
-            data: body
+            data: {
+                image_id: image_id,
+                link: link
+            }
         })
-        res.status(200).json({
-            message: 'Update data success',
+        return res.status(200).json({
             data: tutorials
         })
     } catch (error) {
-        res.status(500).json({
-            message: 'Internal server error',
-            serverMessage: error
+        return res.status(500).json({
+            "error": `${error}`
         })
     }
 }
@@ -82,16 +81,23 @@ const updateTutorial = async (req, res) => {
 const deleteTutorial = async (req, res) => {
     try {
         const {id} = req.params
-        const tutorials = await prisma.tutorials.delete({
+        const tutorials = await prisma.tutorials.findUnique({
             where: {id: Number(id)}
         })
-        res.status(200).json({
+        if(!tutorials){
+            return res.status(404).json({
+                message: 'Data Not Found'
+            })
+        }
+        await prisma.tutorials.delete({
+            where: {id: Number(id)}
+        })
+        return res.status(200).json({
             message: 'Delete data success',
         })
     } catch (error) {
-        res.status(500).json({
-            message: 'Internal server error',
-            serverMessage: error
+        return res.status(500).json({
+            "error": `${error}`
         })
     }
 }
